RM 2527: use instrument name, not inst. frame name for pointings.
These changes look bigger than they actually are.

SpicePointingProvider.java was reworked to take instrument names
on all public methods, rather than ever using the instrument frame
name. A new builder method addInstrumentFrame(instrumentName) allows
calling code to hook up an instrument, with SpicePointingProvider
handling all the frames internally.

IPointingProvider was similarly updated, and everything that called
it, and everything that called everything that called it etc.

In all cases the calling code was defacto passing the instrument
name and not the frame name, so the only impact *should be* that
now the code that calls SPICE will work correctly.